[CIR][CIRGen][NFC] Return scope result in compound stmt builders
Instead of returning a boolean indicating whether the statement was
handled, returns the ReturnExpr of the statement if there is one. It
also adds some extra bookkeeping to ensure that the result is returned
when needed. This allows for better support of GCC's `ExprStmt`
extension.

The logical result was not used: it was handled but it would never fail.
Any errors within builders should likely be handled with asserts and
unreachables since they imply a programmer's error in the code.

ghstack-source-id: d1af9469bb870b0eaba834346952f321cee711ea
Pull Request resolved: https://github.com/llvm/clangir/pull/313